refactor: 일부 클래스 위치 변경
