+++ src/main/java/com/goat/server/global/config/JwtPropertiesConfig.java
@@ -1,6 +1,6 @@
 package com.goat.server.global.config;
 
-import com.goat.server.global.domain.type.JwtProperties;
+import com.goat.server.global.util.jwt.JwtProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
