+++ src/main/java/com/goat/server/global/util/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.goat.server.global.domain.type;
+package com.goat.server.global.util.jwt;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
