+++ src/main/java/com/goat/server/mypage/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.goat.server.mypage.repository;
 
 import com.goat.server.mypage.domain.User;
+import com.goat.server.mypage.dto.JwtUserDetailProjection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
